Finished most of the requirements. Need to add questions to list.

+++ src/edu/yccc/cis174/mmuratake/ExamInterface/Exam.java
@@ -6,6 +6,6 @@ public interface Exam {
 	
 	public List<Question> loadQuestion();
 	
-	public float calculateScore();
+	public double calculateScore();
 
 }
